+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 CHANGELOG
 ---
 
+## 1.2.0
+- Fixed 'isCE' on null
+- Fixed JSON_OBJECT is not a function
+- Fixed 'undefined' error in profile
+
 ## 1.1.0
 - Fixed bugs
 - Added Standard Driver